% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ngrams.R
\name{count_ngrams}
\alias{count_ngrams}
\title{Create ngram frequencies}
\usage{
count_ngrams(ngram_model, by_book = FALSE, doc_title = "title")
}
\arguments{
\item{ngram_model}{A data frame containing an n-gram model}

\item{by_book}{If TRUE, calculate n-gram frequencies by book}

\item{doc_title}{The column name containing document title}
}
\value{
A data frame of n-grams with frequency counts
}
\description{
A function to create frequency counts for an n-gram model
}
\examples{
library(janeaustenr)
library(tidytext)
bigrams <- create_ngrams(austen_books(),"bigram",stopwords=stop_words)
bigram_freqs <- count_ngrams(bigrams)

}
