% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dtm.R
\name{create_dtm}
\alias{create_dtm}
\title{Create a document-term matrix}
\usage{
create_dtm(corpus, stopwords, doc_title = "title")
}
\arguments{
\item{corpus}{A data frame containing columns for title and text}

\item{stopwords}{A data frame containing stop words}

\item{doc_title}{The column name containing document title}
}
\value{
A data frame of a document term matrix.
}
\description{
A function to create a document-term matrix from a corpus
It is a convenience function based around tidytext::cast_dtm
The format of the returned data frame is intended to be suitable as 
input to machine learning tasks
}
\examples{
\dontrun{
library(tidytext)
books <- data.frame(title = c("Book A", "Book B", "Book C"), 
                    text = c("Once upon a time", "A long time ago", 
                    "In a land far away"),
                    stringsAsFactors=FALSE)
book_dtm <- create_dtm(books)
book_dtm_1 <- create_dtm(books,stopwords=stop_words)
}

}
