% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ngrams.R
\name{create_ngrams}
\alias{create_ngrams}
\title{Create ngrams from tidy data}
\usage{
create_ngrams(corpus, ngram_size, stopwords)
}
\arguments{
\item{corpus}{A data frame containing columns for title and text}

\item{ngram_size}{An character string denoting the size of n-gram
(unigram, bigram or trigram)}

\item{stopwords}{A data frame containing stop words}
}
\value{
A data frame of ngrams
}
\description{
A function to create an n-gram model from a tidy data frame.
It is a convenience function based around tidytext functions
}
\examples{
library(janeaustenr)
library(tidytext)
unigrams <- create_ngrams(austen_books(),"unigram")
stopped <- create_ngrams(austen_books(),"unigram",stopwords=stop_words)


}
