% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tfidf.R
\name{create_tfidf}
\alias{create_tfidf}
\title{Calculate tf-idf}
\usage{
create_tfidf(ngram_model, doc_title = "title")
}
\arguments{
\item{ngram_model}{A data frame containing an n-gram ngram_model}

\item{doc_title}{The column name containing document title}
}
\value{
A data frame of n-grams with tf-idf information.
}
\description{
A function to calculate tf-idf for an n-gram ngram_model (stop words not removed).
It is a convenience function based around tidytext::bind_tf_idf
}
\examples{
library(janeaustenr)
library(tidytext)
bigrams <- create_ngrams(austen_books(),"bigram",stopwords=stop_words)
bigram_tfidf <- create_tfidf(bigrams,doc_title="book")

}
